+++ CHANGELOG.md
@@ -7,6 +7,11 @@
     method, provided by `__call()`.
 
 ## 1.5 branch
+### 1.5.10
+* `Exceptionist::__callStatic()` now handles calls starting with "isNot";
+* added `Exceptionist::isFalse()` method;
+* many methods already supported by `Exceptionist` have been documented.
+
 ### 1.5.9
 * `string_ends_with()`, `string_contains()` and `string_starts_with()` are now
     deprecated. Use instead `str_ends_with()`, `str_contains()` and `str_starts_with()`;